+++ ChangeLog.md
@@ -4,12 +4,14 @@ This changelog summarizes changes and fixes which are a part of each revision.  
 where provided, and the git commit history.
 
 * 2017/03/?? - 1.69 - PENDING
-    * 
+    * Fixed issue with nested rows not expanding in concept viewer attached sememes. Also fixed layout problems with nested rows in Chrome.
+    * Fixed layout issues with mapping add fields dialog
 
 * 2017/03/16 - 1.68
     * Mapset include dialogs will now behave as expected if cancel is clicked, by returning to the state they were in prior to any changes. Fixed bug in that code that caused a previously added but cancelled field from being added again. (Jazz: 480730)
     * Added descriptive tooltips and aria-labels to mapset include dialog Change Order and Remove links.
     * Fixed issue with cancelling an edit when the previously displayed content was a Concept Viewer - was not correctly passing view params.
+    * Fixed issue where duplicate fields could be added to a mapset or map items. (Jazz: 480738)
 
 * 2017/03/15 - 1.67
     * Refactored a number of the view params and added Module as a param (Jazz: 461738)